refactor(history): split sidebar responsibilities
